Fixing  bug with World class that was causing teleports to teleport to the block edge, rather than the actual position (which caused people to teleport into walls)

+++ SpongeBridge/src/org/bukkit/World.java
@@ -17,7 +17,7 @@ public class World {
         return this.world.getName();
     }
 
-    public org.spongepowered.api.world.Location getSpongeLocation(int x, int y, int z) throws Exception {
+    public org.spongepowered.api.world.Location getSpongeLocation(double x, double y, double z) throws Exception {
         return this.world.getLocation(x,y,z);
     }
